refactor pages controller
