+++ app/helpers/carts_helper.rb
@@ -1,0 +1,2 @@
+module CartsHelper
+end
