+++ app/helpers/carts_products_helper.rb
@@ -1,0 +1,2 @@
+module CartsProductsHelper
+end
