+++ app/helpers/pages_helper.rb
@@ -1,9 +1,9 @@
 module PagesHelper
   def format_weight(weight)
     if weight < 1000
-      return weight.to_s + " gm"
+      return weight.to_s + " GM"
     end
-    (weight / 1000).to_s + " kg"
+    (weight / 1000).to_s + " KG"
   end
 
   def gst_included_price(price)
