+++ app/models/carts_product.rb
@@ -1,5 +1,4 @@
 class CartsProduct < ApplicationRecord
   belongs_to :product
   belongs_to :cart
-  belongs_to :user
 end
