--- db/migrate/20220707102415_add_reference_to_carts_products.rb
@@ -1,5 +1,0 @@
-class AddReferenceToCartsProducts < ActiveRecord::Migration[7.0]
-  def change
-    add_reference :carts_products, :user, null: false, foreign_key: true
-  end
-end
