+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_07_07_102415) do
+ActiveRecord::Schema[7.0].define(version: 2022_07_07_100026) do
   create_table "carts", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -26,10 +26,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_07_102415) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.integer "product_quantity", default: 1
-    t.integer "user_id", null: false
     t.index ["cart_id"], name: "index_carts_products_on_cart_id"
     t.index ["product_id"], name: "index_carts_products_on_product_id"
-    t.index ["user_id"], name: "index_carts_products_on_user_id"
   end
 
   create_table "orders", force: :cascade do |t|
@@ -64,7 +62,6 @@ ActiveRecord::Schema[7.0].define(version: 2022_07_07_102415) do
   end
 
   add_foreign_key "carts", "users"
-  add_foreign_key "carts_products", "users"
   add_foreign_key "orders", "carts"
   add_foreign_key "orders", "users"
 end
